+++ Source/GateRimSG1/Jaffa/HediffCompProperties_JaffaPrimta.cs
@@ -1,0 +1,12 @@
+using Verse;
+
+namespace GateRimSG1.Jaffa
+{
+    public class HediffCompProperties_JaffaPrimta : HediffCompProperties
+    {
+        public HediffCompProperties_JaffaPrimta()
+        {
+            compClass = typeof(HediffComp_JaffaPrimta);
+        }
+    }
+}